% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting_single_drug.R
\name{FittingSingleDrug}
\alias{FittingSingleDrug}
\title{Fitting single drug dose-response curve}
\usage{
FittingSingleDrug(response.mat, fixed = c(NA, NA, NA, NA),
  nan.handle = c("LL4", "L4"))
}
\arguments{
\item{response.mat}{a matrix with first column as the drug concentrations and second column as the observed responses}

\item{fixed}{a parameter to specify which parameters are fixed and at what value they are fixed. NAs for parameter
that are not fixed.}

\item{nan.handle}{a parameter to specify if L.4 function or LL.4 function is used when fitting with LL.4 produces
NaNs.}
}
\value{
Fitted responses and fitted models are returned.
}
\description{
A function to fit single drug dose-response curve with observed response data
}
\details{
Single drug dose-response curve is fitted with a commonly used 4-paramter log-losistic (4PL) function.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
single.drug.fitted <- FittingSingleDrug(data$dose.response.mats[[1]])
}
\author{
Liye He \email{liye.he@helsinki.fi}
}
\references{
Seber, G. A. F. and Wild, C. J (1989) Nonlinear Regression, New York: Wiley \& Sons (p. 330).
}

