% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathews_screening_data.R
\name{mathews_screening_data}
\alias{mathews_screening_data}
\title{A high-throughput drug combination screening data}
\format{A data frame with the following columns: BlockID, DrugRow, DrugCol, Row, Col, Response, Replicate,
ConcRow, ConcCol, ConcUnit}
\description{
A recent drug combination screening for the treatment of diffuse large B-cell
lymphoma (DLBCL).
}
\references{
Mathews Griner LA, Guha R, Shinn P, Young RM, Keller JM, et al. High-throughput combinatorial
screening identifies drugs that cooperate with ibrutinib to kill activated B-cell-like diffuse large B-cell
lymphoma cells. Proc Natl Acad Sci USA 2014; 111:2349-54.
}

