% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_synergy.R
\name{PlotSynergy}
\alias{PlotSynergy}
\title{Drug interaction landscape}
\usage{
PlotSynergy(data, type = "2D", save.file = FALSE, pair.index = NULL,
  legend.start = NULL, legend.end = NULL, x.range = NULL,
  y.range = NULL)
}
\arguments{
\item{data}{a list object generated by function \code{\link{CalculateSynergy}}.}

\item{type}{a parameter to specify the type of the interaction landscape, 2D, 3D
or both. By default, 2D interaction landscape is returned.}

\item{save.file}{a logical parameter to specify if the interaction landscape is saved as a pdf file
in the current working directory or returned as an R object. By default, it is FALSE.}

\item{pair.index}{a parameter to specify which drug combination if there are many drug combinations in the data.
By default, it is NULL so that the synergy score visualization of all the drug combinations in the data is returned.}

\item{legend.start}{a parameter to specify the starting point of the legend. By defualt, it is NULL so the legend 
starting point is fixed by the data automatically.}

\item{legend.end}{a parameter to specify the ending point of the legend. By defualt, it is NULL so the legend 
ending point is fixed by the data automatically.}

\item{x.range}{a parameter to specify the starting and ending concentration of the drug on x-axis. Use e.g., c(1, 3) to 
specify that only from 1st to 3rd concentrations of the drug on x-axis are used. By default, it is NULl so all the 
concentrations are used.}

\item{y.range}{a parameter to specify the starting and ending concentration of the drug on y-axis. Use e.g., c(1, 3) to 
specify that only from 1st to 3rd concentrations of the drug on y-axis are used. By default, it is NULl so all the 
concentrations are used.}
}
\value{
a pdf file or the interaction landscapes are only displayed depending on
the save.file parameter.
}
\description{
A function to visualize the synergy scores for drug combinations as 2D or 3D
interaction landscape over the dose-response matrix.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
scores <- CalculateSynergy(data)
PlotSynergy(scores, "2D")
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

