% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP.R
\name{ZIP}
\alias{ZIP}
\title{Delta synergy score based on zero interaction potency (ZIP) model}
\usage{
ZIP(response.mat, correction = TRUE, Emin = 0, Emax = 100,
  nan.handle = c("LL4", "L4"))
}
\arguments{
\item{response.mat}{a dose-response matrix with concentrations as row names and column names}

\item{correction}{a parameter to specify if the baseline correction is used or not. Defaults to TRUE.}

\item{Emin}{the minimal effect of the drug used in the 4-parameter log-logistic function to fit the dose-response 
curve. If it is not NA, it is fixed the value assigned by the user. Defaults to 0.}

\item{Emax}{the maximal effect of the drug used in the 4-parameter log-logistic function to fit the dose-response 
curve. If it is not NA, it is fixed the value assigned by the user. Defaults to 100.}

\item{nan.handle}{a parameter to specify if L.4 function or LL.4 function is used when fitting with LL.4 produces
NaNs.}
}
\value{
A matrix of delta scores for all the dose pairs for a drug combination. For a does pair with at least one zero concentration, 0 is used as the synergy score.
}
\description{
A function to calculate delta synergy score based on zero interaction potency (ZIP) model
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
delta.score <- ZIP(data$dose.response.mats[[1]])
}
\author{
Liye He \email{liye.he@helsinki.fi}, Jing Tang \email{jing.tang@helsinki.fi}
}
\references{
Yadav B, Wennerberg K, Aittokallio T, Tang J. Searching for Drug Synergy in Complex Dose-Response Landscape Using an Interaction Potency Model.
Computational and Structural Biotechnology Journal 2015; 13: 504-513.
}

